Abort Check

A wrapper on err() that sets the subclass to be 'chk_error'.

abort_chk(..., n = NULL, tidy = TRUE)

Arguments

...
Multiple objects that are converted to a string using paste0(..., collapse = '').

n
The value of n for converting sprintf-like types.

tidy
A flag specifying whether capitalize the first character and add a missing period.

Value

Throws an error of class 'chk_error'.

Details

It is exported to allow users to easily construct their own chk_ functions.

See also

Examples

try(abort_chk("x must be NULL"))
#> Error : X must be NULL.
try(abort_chk("`x` must be NULL"))
#> Error : `x` must be NULL.
try(abort_chk("there %r %n problem value%s", n = 1))
#> Error : There is 1 problem value.
try(abort_chk("there %r %n problem value%s", n = 1.5))
#> Error : There are 1.5 problem values.