Check Names

Checks the names of an object.

Usage

check_names(
  x,
  names = character(0),
  exclusive = FALSE,
  order = FALSE,
  x_name = NULL
)

Arguments

x

The object to check.

names

A character vector of the required names.

exclusive

A flag specifying whether x must only contain the required names.

order

A flag specifying whether the order of the required names in x must match the order in names.

x_name

A string of the name of object x or NULL.

Value

An informative error if the test fails or an invisible copy of x.

See also

Examples

x <- c(x = 1, y = 2)
check_names(x, c("y", "x"))
try(check_names(x, c("y", "x"), order = TRUE))
#> Error in eval(expr, envir) : 
#>   `names(x)` must include 'y' and 'x' in that order.
try(check_names(x, "x", exclusive = TRUE))
#> Error in eval(expr, envir) : `names(x)` must not include 'y'.