Check Files or Directories Exist

Checks if files or directorys exist.

chk_file(x, err = TRUE)

chk_dir(x, err = TRUE)

chk_dirs(x, err = TRUE)

chk_files(x, err = TRUE)

Arguments

x
The object to check.

err
A flag specifying whether to generate an error message if the check fails.

Value

TRUE if passes check. Otherwise if throws an informative error unless err = FALSE in which case it returns FALSE.

Details

chk_file(): Check Files Exist

Checks if files exist using:

all(file.exists(x) && !dir.exists(x))

Functions

chk_dir: Check Directories Exist

Checks if directories exist using:

all(dir.exists(x))

chk_dirs: Check Directories Exist (DEPRECATED)

Checks if directories exist using:

all(dir.exists(x))

chk_files: Check Files Exist (DEPRECATED)

Checks if files exist using:

all(file.exists(x) && !dir.exists(x))

Examples

# chk_file
try(chk_file(tempfile()))

#> Error : Can't find file '/var/folders/48/q6ltldjs251000_wvjrdy_vm0000gn/T//RtmphzZlIu/file33e16b435ddb'.

# chk_dir
try(chk_dir(tempfile()))

#> Error : Can't find directory '/var/folders/48/q6ltldjs251000_wvjrdy_vm0000gn/T//RtmphzZlIu/file33e113c70be1'.