Check/Validate Is

Checks if is a particular type of object.

chk_is(x, class, x_name = NULL)

vld_is(x, class)

chk_whole_numeric(x, x_name = NULL)

vld_whole_numeric(x)

chk_list(x, x_name = NULL)

vld_list(x)

chk_function(x, x_name = NULL)

vld_function(x)

Arguments

x
The object to check.

class
A string specifying the class.

x_name
A string of the name of object x or NULL.

Value

The chk_ functions throw an informative error if the test fails. The vld_ functions return a flag indicating whether the test was met.

Functions

chk_is: Check Is

Checks if inherits from class using vld_is().

vld_is: Validate Is

Validates inherits from class using

inherits(x, class)

Class should be a string.

chk_whole_numeric: Check Whole Numeric

Checks if integer vector or double equivalent using vld_whole_numeric().

The chk_whole_number() function checks if non-missing integer scalar or double equivalent.

vld_whole_numeric: Validate Whole Numeric

Validates integer vector or double equivalent using

is.integer(x) || (is.double(x) && isTRUE(all.equal(x, as.integer(x))))

chk_list: Check List

Checks if is a list using vld_list().

vld_list: Validate List

Validates is a list using

is.list(x)

chk_function: Check Function

Checks if is a function using vld_function().

vld_function: Validate Function

Validates is a function using:

is.function(x)

Examples

# chk_is
chk_is(1, "numeric")
try(chk_is(1, "character"))

#> Error : `1` must inherit from class 'character'.
#> Backtrace:
#>      █
#>   1. ├─base::tryCatch(...)
#>   2. │ └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>   3. │   ├─base:::tryCatchOne(...)
#>   4. │   │ └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>   5. │   └─base:::tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#>   6. │     └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>   7. │       └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>   8. ├─base::withCallingHandlers(...)
#>   9. ├─base::saveRDS(...)
#>  10. ├─base::do.call(...)
#>  11. ├─(function (what, args, quote = FALSE, envir = parent.frame()) ...
#>  12. └─(function (..., crayon_enabled, crayon_colors, pkgdown_internet) ...
#>  13.   └─pkgdown::build_site(...)
#>  14.     └─pkgdown:::build_site_local(...)
#>  15.       └─pkgdown::build_reference(...)
#>  16.         └─purrr::map(...)
#>  17.           └─pkgdown:::.f(.x[[i]], ...)
#>  18.             └─pkgdown:::data_reference_topic(...)
#>  19.               └─pkgdown:::run_examples(...)
#>  20.                 └─pkgdown:::highlight_examples(code, topic, env = env)
#>  21.                   └─evaluate::evaluate(x, child_env(env), new_device = TRUE)
#>  22.                     └─evaluate:::evaluate_call(...)
#>  23.                       ├─evaluate:::timing_fn(...)
#>  24.                       ├─evaluate:::handle(...)
#>  25.                       │ └─base::try(f, silent = TRUE)
#>  26.                       │   └─base::tryCatch(...)
#>  27.                       │     └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>  28.                       │       └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>  29.                       │         └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>  30.                       ├─base::withCallingHandlers(...)
#>  31.                       ├─base::withVisible(eval(expr, envir, enclos))
#>  32.                       └─base::eval(expr, envir, enclos)
#>  33.                         └─base::eval(expr, envir, enclos)
#>  34.                           ├─base::try(chk_is(1, "character"))
#>  35.                           │ └─base::tryCatch(...)
#>  36.                           │   └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>  37.                           │     └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>  38.                           │       └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>  39.                           └─chk::chk_is(1, "character")
#>  40.                             └─chk:::.abort(x_name, " must inherit from class '", class, "'.") /home/kirill/git/R/chk/R/chk-is.R:26:2

# vld_is
vld_is(numeric(0), "numeric")

#> [1] TRUE

vld_is(1, "numeric")

#> [1] TRUE

vld_is(NA_real_, "numeric")

#> [1] TRUE

vld_is(1, "character")

#> [1] FALSE

vld_is(NULL, "numeric")

#> [1] FALSE

# chk_whole_numeric
chk_whole_numeric(1)
try(chk_whole_numeric(1.1))

#> Error : `1.1` must be a whole numeric vector (integer vector or double equivalent).
#> Backtrace:
#>      █
#>   1. ├─base::tryCatch(...)
#>   2. │ └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>   3. │   ├─base:::tryCatchOne(...)
#>   4. │   │ └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>   5. │   └─base:::tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#>   6. │     └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>   7. │       └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>   8. ├─base::withCallingHandlers(...)
#>   9. ├─base::saveRDS(...)
#>  10. ├─base::do.call(...)
#>  11. ├─(function (what, args, quote = FALSE, envir = parent.frame()) ...
#>  12. └─(function (..., crayon_enabled, crayon_colors, pkgdown_internet) ...
#>  13.   └─pkgdown::build_site(...)
#>  14.     └─pkgdown:::build_site_local(...)
#>  15.       └─pkgdown::build_reference(...)
#>  16.         └─purrr::map(...)
#>  17.           └─pkgdown:::.f(.x[[i]], ...)
#>  18.             └─pkgdown:::data_reference_topic(...)
#>  19.               └─pkgdown:::run_examples(...)
#>  20.                 └─pkgdown:::highlight_examples(code, topic, env = env)
#>  21.                   └─evaluate::evaluate(x, child_env(env), new_device = TRUE)
#>  22.                     └─evaluate:::evaluate_call(...)
#>  23.                       ├─evaluate:::timing_fn(...)
#>  24.                       ├─evaluate:::handle(...)
#>  25.                       │ └─base::try(f, silent = TRUE)
#>  26.                       │   └─base::tryCatch(...)
#>  27.                       │     └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>  28.                       │       └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>  29.                       │         └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>  30.                       ├─base::withCallingHandlers(...)
#>  31.                       ├─base::withVisible(eval(expr, envir, enclos))
#>  32.                       └─base::eval(expr, envir, enclos)
#>  33.                         └─base::eval(expr, envir, enclos)
#>  34.                           ├─base::try(chk_whole_numeric(1.1))
#>  35.                           │ └─base::tryCatch(...)
#>  36.                           │   └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>  37.                           │     └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>  38.                           │       └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>  39.                           └─chk::chk_whole_numeric(1.1)
#>  40.                             └─chk:::.abort(x_name, " must be a whole numeric vector (integer vector or double equivalent).") /home/kirill/git/R/chk/R/chk-is.R:66:2

# vld_whole_numeric
vld_whole_numeric(1)

#> [1] TRUE

vld_whole_numeric(NA_real_)

#> [1] TRUE

vld_whole_numeric(1:2)

#> [1] TRUE

vld_whole_numeric(double(0))

#> [1] TRUE

vld_whole_numeric(TRUE)

#> [1] FALSE

vld_whole_numeric(1.5)

#> [1] FALSE

# chk_list
chk_list(list())
try(chk_list(1))

#> Error : `1` must be a list.
#> Backtrace:
#>      █
#>   1. ├─base::tryCatch(...)
#>   2. │ └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>   3. │   ├─base:::tryCatchOne(...)
#>   4. │   │ └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>   5. │   └─base:::tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#>   6. │     └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>   7. │       └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>   8. ├─base::withCallingHandlers(...)
#>   9. ├─base::saveRDS(...)
#>  10. ├─base::do.call(...)
#>  11. ├─(function (what, args, quote = FALSE, envir = parent.frame()) ...
#>  12. └─(function (..., crayon_enabled, crayon_colors, pkgdown_internet) ...
#>  13.   └─pkgdown::build_site(...)
#>  14.     └─pkgdown:::build_site_local(...)
#>  15.       └─pkgdown::build_reference(...)
#>  16.         └─purrr::map(...)
#>  17.           └─pkgdown:::.f(.x[[i]], ...)
#>  18.             └─pkgdown:::data_reference_topic(...)
#>  19.               └─pkgdown:::run_examples(...)
#>  20.                 └─pkgdown:::highlight_examples(code, topic, env = env)
#>  21.                   └─evaluate::evaluate(x, child_env(env), new_device = TRUE)
#>  22.                     └─evaluate:::evaluate_call(...)
#>  23.                       ├─evaluate:::timing_fn(...)
#>  24.                       ├─evaluate:::handle(...)
#>  25.                       │ └─base::try(f, silent = TRUE)
#>  26.                       │   └─base::tryCatch(...)
#>  27.                       │     └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>  28.                       │       └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>  29.                       │         └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>  30.                       ├─base::withCallingHandlers(...)
#>  31.                       ├─base::withVisible(eval(expr, envir, enclos))
#>  32.                       └─base::eval(expr, envir, enclos)
#>  33.                         └─base::eval(expr, envir, enclos)
#>  34.                           ├─base::try(chk_list(1))
#>  35.                           │ └─base::tryCatch(...)
#>  36.                           │   └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>  37.                           │     └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>  38.                           │       └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>  39.                           └─chk::chk_list(1)
#>  40.                             └─chk:::.abort(x_name, " must be a list.") /home/kirill/git/R/chk/R/chk-is.R:106:2

# vld_list
vld_list(list())

#> [1] TRUE

vld_list(list(x = 1))

#> [1] TRUE

vld_list(mtcars)

#> [1] TRUE

vld_list(1)

#> [1] FALSE

vld_list(NULL)

#> [1] FALSE

# chk_function
chk_function(mean)
try(chk_function(1))

#> Error : `1` must be a function.
#> Backtrace:
#>      █
#>   1. ├─base::tryCatch(...)
#>   2. │ └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>   3. │   ├─base:::tryCatchOne(...)
#>   4. │   │ └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>   5. │   └─base:::tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#>   6. │     └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>   7. │       └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>   8. ├─base::withCallingHandlers(...)
#>   9. ├─base::saveRDS(...)
#>  10. ├─base::do.call(...)
#>  11. ├─(function (what, args, quote = FALSE, envir = parent.frame()) ...
#>  12. └─(function (..., crayon_enabled, crayon_colors, pkgdown_internet) ...
#>  13.   └─pkgdown::build_site(...)
#>  14.     └─pkgdown:::build_site_local(...)
#>  15.       └─pkgdown::build_reference(...)
#>  16.         └─purrr::map(...)
#>  17.           └─pkgdown:::.f(.x[[i]], ...)
#>  18.             └─pkgdown:::data_reference_topic(...)
#>  19.               └─pkgdown:::run_examples(...)
#>  20.                 └─pkgdown:::highlight_examples(code, topic, env = env)
#>  21.                   └─evaluate::evaluate(x, child_env(env), new_device = TRUE)
#>  22.                     └─evaluate:::evaluate_call(...)
#>  23.                       ├─evaluate:::timing_fn(...)
#>  24.                       ├─evaluate:::handle(...)
#>  25.                       │ └─base::try(f, silent = TRUE)
#>  26.                       │   └─base::tryCatch(...)
#>  27.                       │     └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>  28.                       │       └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>  29.                       │         └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>  30.                       ├─base::withCallingHandlers(...)
#>  31.                       ├─base::withVisible(eval(expr, envir, enclos))
#>  32.                       └─base::eval(expr, envir, enclos)
#>  33.                         └─base::eval(expr, envir, enclos)
#>  34.                           ├─base::try(chk_function(1))
#>  35.                           │ └─base::tryCatch(...)
#>  36.                           │   └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>  37.                           │     └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>  38.                           │       └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>  39.                           └─chk::chk_function(1)
#>  40.                             └─chk:::.abort(x_name, " must be a function.") /home/kirill/git/R/chk/R/chk-is.R:141:2

# vld_function
vld_function(mean)

#> [1] TRUE

vld_function(function(x) x)

#> [1] TRUE

vld_function(1)

#> [1] FALSE

vld_function(list(1))

#> [1] FALSE