Miscellaneous Checks

Miscellaneous checks.

chk_length(x, length = 1L, err = TRUE, x_name = NULL)

chk_no_missing(x, err = TRUE, x_name = NULL)

chk_unique(x, incomparables = FALSE, err = TRUE, x_name = NULL)

chk_named(x, err = TRUE, x_name = NULL)

chk_in(x, values = c(0, 1, NA), err = TRUE, x_name = NULL)

chk_match(x, regexp = ".+", err = TRUE, x_name = NULL)

Arguments

x
The object to check.

length
A count of the length.

err
A flag specifying whether to generate an error message if the check fails.

x_name
A string of the name of object x or NULL.

incomparables
A vector of values that cannot be compared. FALSE means that all values can be compared.

values
A vector of the permitted values.

regexp
A string of a regular expression.

Value

TRUE if passes check. Otherwise if throws an informative error unless err = FALSE in which case it returns FALSE.

Details

chk_length(): Check Length

Checks if is length length using:

length(x) == length

length should be a count.

Functions

chk_no_missing: Check No Missing Values

Checks if no missing values using:

!anyNA(x)

chk_unique: Check Unique

Checks if unique using:

!anyDuplicated(x, incomparables = incomparables)

chk_named: Check Named

Checks if is named using:

!is.null(names(x))

chk_in: Check In

Checks if all values in values using equivalent of:

all(match(x, values, nomatch = 0) > 0)

chk_match: Check Matches

Checks if all values match regular expression using:

all(grepl(regexp, x))

regexp should be a non-missing character scalar.

Examples

# chk_length
try(chk_length(numeric(0)))

#> Error : `numeric(0)` must be length 1, not 0.

try(chk_length(1:2))

#> Error : `1:2` must be length 1, not 2.

chk_length(1)

#> [1] TRUE

chk_length(NA_character_)

#> [1] TRUE

chk_length(Sys.Date())

#> [1] TRUE

chk_length(Sys.time())

#> [1] TRUE

chk_length(1:2, length = 2)

#> [1] TRUE

# chk_no_missing
try(chk_no_missing(NA))

#> Error : `NA` must not have missing values.

try(chk_no_missing(c(1, NA)))

#> Error : `c(1, NA)` must not have missing values.

chk_no_missing(character(0))

#> [1] TRUE

chk_no_missing(1)

#> [1] TRUE

chk_no_missing(1:10)

#> [1] TRUE

# chk_unique
chk_unique(NULL)

#> [1] TRUE

chk_unique(numeric(0))

#> [1] TRUE

chk_unique(c(NA, 2))

#> [1] TRUE

try(chk_unique(c(NA, NA, 2)))

#> Error : `c(NA, NA, 2)` must be unique.

chk_unique(c(NA, NA, 2), incomparables = NA)

#> [1] TRUE

# chk_named
try(chk_named(1))

#> Error : `1` must be named.

try(chk_named(list(1)))

#> Error : `list(1)` must be named.

chk_named(c(x = 1))

#> [1] TRUE

chk_named(list(x = 1))

#> [1] TRUE

chk_named(c(x = 1)[-1])

#> [1] TRUE

chk_named(list(x = 1)[-1])

#> [1] TRUE

# chk_in
chk_in(numeric(0), 1:10)

#> [1] TRUE

chk_in(1, 1:10)

#> [1] TRUE

try(chk_in(11, 1:10))

#> Error : `11` must match 1, 2, 3, 4, 5, 6, 7, 8, 9 or 10, not 11.

# chk_match
try(chk_match(""))

#> Error : `""` must match regular expression '.+'.

chk_match("1")

#> [1] TRUE

try(chk_match("1", regexp = "2"))

#> Error : `"1"` must match regular expression '2'.

try(chk_match(NA_character_, regexp = ".*"))

#> Error : `NA_character_` must match regular expression '.*'.

chk_match(c("a"), regexp = ".*")

#> [1] TRUE