Check Object

Efficiently checks object.

chk_true(x, err = TRUE)

chk_false(x, err = TRUE)

chk_flag(x, err = TRUE)

chk_lgl(x, err = TRUE)

chk_number(x, err = TRUE)

chk_proportion(x, err = TRUE)

chk_whole_number(x, err = TRUE)

chk_count(x, err = TRUE)

chk_string(x, err = TRUE)

chk_date(x, err = TRUE)

chk_datetime(x, err = TRUE)

chk_scalar(x, mode = "any", err = TRUE)

chk_vector(x, mode = "any", err = TRUE)

chk_atomic(x, err = TRUE)

chk_whole_numeric(x, err = TRUE)

chk_grepl(x, pattern = ".+", err = TRUE)

chk_no_missing(x, err = TRUE)

chk_unique(x, err = TRUE)

chk_null(x, err = TRUE)

chk_not_null(x, err = TRUE)

chk_named(x, err = TRUE)

chk_unused(..., err = TRUE)

chk_used(..., err = TRUE)

chk_function(x, err = TRUE)

chk_is(x, class, err = TRUE)

chk_identical(x, y, err = TRUE)

chk_equal(x, y, tolerance = sqrt(.Machine$double.eps), err = TRUE)

chk_equivalent(x, y, tolerance = sqrt(.Machine$double.eps), err = TRUE)

chk_lt(x, value = 0, err = TRUE)

chk_lte(x, value = 0, err = TRUE)

chk_gt(x, value = 0, err = TRUE)

chk_gte(x, value = 0, err = TRUE)

chk_range(x, range = c(0, Inf), err = TRUE)

chk_match(x, values = c(0, Inf), err = TRUE)

chk_dir(x, err = TRUE)

chk_dirs(x, err = TRUE)

chk_file(x, err = TRUE)

chk_files(x, err = TRUE)

Arguments

x
The object to check.

err
A flag specifying whether to generate an error message if the check fails.

mode
A string naming an atomic mode or "list" or "expression" or "any".

pattern
A string of a regular expression.

...
Additional arguments.

class
A string specifying the class.

y
An object to check against.

tolerance
A non-negative numeric scalar.

value
A non-missing scalar of a value.

range
A vector of length 2 of the lower and upper permitted values.

values
A vector of the permitted values.

Value

TRUE if passes check. Otherwise if throws an informative error unless err = FALSE in which case it returns FALSE.

Details

chk_true(): Check TRUE

Checks if TRUE using:

isTRUE(x)

Functions

chk_false: Check FALSE

Checks if FALSE using:

isFALSE(x)

chk_flag: Check Flag

Checks if non-missing logical scalar using:

is.logical(x) && length(x) == 1L && !anyNA(x)

chk_lgl: Check Lgl

Checks if logical scalar using:

is.logical(x) && length(x) == 1L && !anyNA(x)

chk_number: Check Number

Checks if non-mising numeric scalar using:

is.numeric(x) && length(x) == 1L && !anyNA(x)

chk_proportion: Check Proportion

Checks if non-mising numeric scalar between 0 and 1 using:

is.numeric(x) && length(x) == 1L && !anyNA(x) && x >= 0 && x <= 1

chk_whole_number: Check Whole Number

Checks if non-mising integer scalar or double equivalent using:

is.numeric(x) && length(x) == 1L && !anyNA(x) && (is.integer(x) || isTRUE(all.equal(x, as.integer(x))))

chk_count: Check Count

Checks if non-missing non-negative integer scalar or double equivalent using:

is.numeric(x) && length(x) == 1L && !anyNA(x) && x >= 0 && isTRUE(all.equal(x, as.integer(x)))

chk_string: Check String

Checks if non-missing character scalar using:

is.character(x) && length(x) == 1L && !anyNA(x)

chk_date: Check String

Checks if non-missing Date scalar using:

inherits(x, "Date") && length(x) == 1L && !anyNA(x)

chk_datetime: Check DateTime

Checks if non-missing POSIXct scalar using:

inherits(x, "POSIXct") && length(x) == 1L && !anyNA(x)

chk_scalar: Check Scalar

Checks if scalar using:

is.vector(x, mode = mode) && length(x) == 1L

mode should be a string.

chk_vector: Check Vector

Checks if vector using:

is.vector(x, mode = mode)

mode should be a string.

chk_atomic: Check Atomic

Checks if atomic using:

is.atomic(x)

chk_whole_numeric: Check Whole Numeric

Checks if integer vector or double equivalent using the equivalent of:

is.integer(x) || (is.double(x) && isTRUE(all.equal(x, as.integer(x))))

chk_grepl: Check Matches Regular Expression

Checks if all values match regular expression pattern using:

all(grepl(pattern, x))

pattern should be a non-missing character scalar.

chk_no_missing: Check No Missing Values

Checks if no missing values using:

!anyNA(x)

chk_unique: Check Unique

Checks if unique using:

!anyDuplicated(x)

chk_null: Check NULL

Checks if NULL using:

is.null(x)

chk_not_null: Check Not NULL

Checks if not NULL using:

!is.null(x)

chk_named: Check Named

Checks if is named using:

!is.null(names(x))

chk_unused: Check ... Unused

Checks if is ... unused using:

!length(list(...))

chk_used: Check ... Used

Checks if is ... used using:

length(list(...))

chk_function: Check Function

Checks if is a function using:

is.function(x)

chk_is: Check Is

Checks if inherits from class using:

inherits(x, class)

Class should be a string.

chk_identical: Check Identical

Checks if is identical to y using:

identical(x, y)

chk_equal: Check Equal

Checks if is equal (identical within tolerance) to y using:

isTRUE(all.equal(x, y, tolerance))

chk_equivalent: Check Equivalent

Checks if is equivalent (equal ignoring attributes) to y using:

isTRUE(all.equal(x, y, tolerance, check.attributes = FALSE))

chk_lt: Check Less Than

Checks if all non-missing values are less than value using:

all(x[!is.na(x)] < value))

value should be a non-missing scalar.

chk_lte: Check Less Than or Equal To

Checks if all non-missing values are less than or equal to y using:

all(x[!is.na(x)] <= value))

value should be a non-missing scalar.

chk_gt: Check Greater Than

Checks if all non-missing values are greater than value using:

all(x[!is.na(x)] > value))

value should be a non-missing scalar.

chk_gte: Check Greater Than or Equal To

Checks if all non-missing values are greater than or equal to y using:

all(x[!is.na(x)] >= value))

value should be a non-missing scalar.

chk_range: Check Range

Checks if all non-missing values fall within range using:

all(x[!is.na(x)] >= range[1] & x[!is.na(x)] <= range[2])

Range should be a non-mising sorted vector of length 2.

chk_match: Check Match

Checks if all values match values using equivalent of:

all(match(x, values, nomatch = 0) > 0)

chk_dir: Check Directory Exists

Checks if directory exist using:

length(x) == 1L && dir.exists(x))

chk_dirs: Check Directories Exist

Checks if directories exist using:

all(dir.exists(x))

chk_file: Check File Exists

Checks if file exists using:

length(x) == 1L & file.exists(x) && !dir.exists(x)

chk_files: Check Files Exist

Checks if files exist using:

all(file.exists(x) && !dir.exists(x))

Examples

# chk_true
chk_true(1, err = FALSE)

#> [1] FALSE

chk_true(NA, err = FALSE)

#> [1] FALSE

chk_true(FALSE, err = FALSE)

#> [1] FALSE

chk_true(TRUE)

#> [1] TRUE

chk_true(c(TRUE, TRUE), err = FALSE)

#> [1] FALSE

# chk_false
chk_false(0, err = FALSE)

#> [1] FALSE

chk_false(NA, err = FALSE)

#> [1] FALSE

chk_false(TRUE, err = FALSE)

#> [1] FALSE

chk_false(FALSE)

#> [1] TRUE

chk_false(c(FALSE, FALSE), err = FALSE)

#> [1] FALSE

# chk_flag
chk_flag(1, err = FALSE)

#> [1] FALSE

chk_flag(NA, err = FALSE)

#> [1] FALSE

chk_flag(TRUE)

#> [1] TRUE

chk_flag(FALSE)

#> [1] TRUE

chk_flag(c(TRUE, TRUE), err = FALSE)

#> [1] FALSE

# chk_lgl
chk_lgl(1, err = FALSE)

#> [1] FALSE

chk_lgl(NA)

#> [1] TRUE

chk_lgl(TRUE)

#> [1] TRUE

chk_lgl(FALSE)

#> [1] TRUE

chk_lgl(c(TRUE, TRUE), err = FALSE)

#> [1] FALSE

# chk_unique
chk_unique(NA)

#> [1] TRUE

chk_unique(c(NA, NA), err = FALSE)

#> [1] FALSE