Check Vector

Checks if is a vector using

(is.atomic(x) && !is.matrix(x) && !is.array(x)) || is.list(x)

Usage

chk_vector(x, x_name = NULL)

vld_vector(x)

Arguments

x

The object to check.

x_name

A string of the name of object x or NULL.

Value

The chk_ function throws an informative error if the test fails or returns the original object if successful so it can used in pipes.

The vld_ function returns a flag indicating whether the test was met.

Details

is.vector(x) is not reliable because it returns TRUE only if the object is a vector with no attributes apart from names.

Functions

vld_vector(): Validate Vector

See also

Examples

# chk_vector
chk_vector(1)
chk_vector(list())
try(chk_vector(matrix(1)))
#> Error in eval(expr, envir) : `matrix(1)` must be a vector.
# vld_vector
vld_vector(1)
#> [1] TRUE