Stop, Warning and Message

Wrappers on abort(), warn() and inform().

err(..., n = NULL, .subclass = NULL)

wrn(..., n = NULL, .subclass = NULL)

msg(..., n = NULL, .subclass = NULL)

Arguments

...
zero or more objects which can be coerced to character (and which are pasted together with no separator) or a single condition object.

n
The value of n for converting sprintf-like types.

.subclass
Subclass of the condition. This allows your users to selectively handle the conditions signalled by your functions.

Functions

err: Error

wrn: Warning

msg: Message

sprintf-like types

The following sprintf-like types can be used in the custom messages:

n

The value of n.

s

'' if n == 1 otherwise 's'

r

'is' if n == 1 otherwise 'are'

Examples

# err
try(err("this is an error"))

#> Error : this is an error
#> Backtrace:
#>      █
#>   1. ├─base::tryCatch(...)
#>   2. │ └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>   3. │   ├─base:::tryCatchOne(...)
#>   4. │   │ └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>   5. │   └─base:::tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#>   6. │     └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>   7. │       └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>   8. ├─base::withCallingHandlers(...)
#>   9. ├─base::saveRDS(...)
#>  10. ├─base::do.call(...)
#>  11. ├─(function (what, args, quote = FALSE, envir = parent.frame()) ...
#>  12. └─(function (..., crayon_enabled, crayon_colors, pkgdown_internet) ...
#>  13.   └─pkgdown::build_site(...)
#>  14.     └─pkgdown:::build_site_local(...)
#>  15.       └─pkgdown::build_reference(...)
#>  16.         └─purrr::map(...)
#>  17.           └─pkgdown:::.f(.x[[i]], ...)
#>  18.             └─pkgdown:::data_reference_topic(...)
#>  19.               └─pkgdown:::run_examples(...)
#>  20.                 └─pkgdown:::highlight_examples(code, topic, env = env)
#>  21.                   └─evaluate::evaluate(x, child_env(env), new_device = TRUE)
#>  22.                     └─evaluate:::evaluate_call(...)
#>  23.                       ├─evaluate:::timing_fn(...)
#>  24.                       ├─evaluate:::handle(...)
#>  25.                       │ └─base::try(f, silent = TRUE)
#>  26.                       │   └─base::tryCatch(...)
#>  27.                       │     └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>  28.                       │       └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>  29.                       │         └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>  30.                       ├─base::withCallingHandlers(...)
#>  31.                       ├─base::withVisible(eval(expr, envir, enclos))
#>  32.                       └─base::eval(expr, envir, enclos)
#>  33.                         └─base::eval(expr, envir, enclos)
#>  34.                           ├─base::try(err("this is an error"))
#>  35.                           │ └─base::tryCatch(...)
#>  36.                           │   └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>  37.                           │     └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>  38.                           │       └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>  39.                           └─chk::err("this is an error")

# wrn
wrn("this is a warning")

#> Warning: this is a warning

wrn("there %r %n problem value%s")

#> Warning: there %r %n problem value%s

wrn("there %r %n problem value%s", n = 1)

#> Warning: there is 1 problem value

wrn("there %r %n problem value%s", n = 2)

#> Warning: there are 2 problem values

# msg
msg("this is a message")

#> this is a message

msg("there %r %n problem value%s")

#> there %r %n problem value%s

msg("there %r %n problem value%s", n = 1)

#> there is 1 problem value

msg("there %r %n problem value%s", n = 2)

#> there are 2 problem values